+++ index.ts
@@ -1,0 +1,5 @@
+async function request(url: string) {
+    return (await fetch(url, {}))
+}
+
+console.log(await request())